pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Serggi0 <serve4mia@gmail.com>"]
packages = [
  { include = "page_loader" },
]

[tool.poetry.dependencies]
python = "^3.8"
lxml = "^4.6.3"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest-cov = "^2.12.0"
requests = "^2.25.1"
coverage = "^5.5"
Pillow = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"